v1.4.9
compatible for raspberrypi.

+++ slscore/SLSLog.cpp
@@ -84,7 +84,7 @@ void CSLSLog::print_log(int level, const char *fmt, va_list vl)
     sls_gettime_fmt(cur_time, cur_time_sec, "%Y-%m-%d %H:%M:%S");
     vsnprintf (buf , 4095 , fmt , vl);
     //sprintf(buf_info, "%s %s: %s\n" , cur_time, LOG_LEVEL_NAME[level], buf);
-    sprintf(buf_info, "%s:%03d %s %s: %s\n" , cur_time, cur_time_msec, APP_NAME, LOG_LEVEL_NAME[level], buf);
+    sprintf(buf_info, "%s:%03d %s %s: %s\n" , cur_time, (int)cur_time_msec, APP_NAME, LOG_LEVEL_NAME[level], buf);
     printf(buf_info);
 
     if (m_log_file) {
